+++ src/connection/esb.c
@@ -1,6 +1,6 @@
-#include "globals.h"
-#include "system.h"
-#include "hid.h"
+#include "../globals.h"
+#include "../system/system.h"
+#include "../hid.h"
 
 #include <zephyr/drivers/clock_control/nrf_clock_control.h>
 #include <zephyr/sys/crc.h>
